fix(guest): run the guest environment setup on the universal-initramfs path
The universal initramfs made the agent PID 1 and replaced mvm-oci-init, but
only mvm-oci-init's mounting was carried over. Everything else it did was
silently dropped from the default boot path: loopback bring-up, the resolver
seed, the egress client, the egress CA, the verb grant, netinit, and the
mediated-tool substitution. The host-signer anchor was the one step ported,
because it alone failed the boot loudly.

The result was a workload that booted fine and could not reach anything. It
still received all eight proxy environment variables, pointing at a loopback
port with nothing behind it, so every egress attempt failed as
`Network unreachable` rather than as a policy denial.

Extract the steps into mvm-agentd's guest_bootstrap so both inits share one
copy — two inits drifting is how the steps went missing in the first place.
apply_activation runs them after the pivot, since they write into the
workload root, and before the privilege drop, since they need root. A
required-but-unresolvable egress client now aborts activation instead of
booting a workload that cannot reach its admitted egress.

Wiring the substitution up exposed a bug in it: MS_BIND resolves its target,
and /bin/ping on a busybox image is a symlink to /bin/busybox, so the mount
replaced every applet including sh. It had never run on this path, so it had
never been reached. replace_symlink_target swaps the link for a plain file
first and declines the substitution when it cannot, which is what happens on
a verity-sealed rootfs.

Verified live on macOS/HVF with `machine run --image alpine`: wget over https
returns the page, lo is up, /run/mvm is provisioned, and ICMP over vsock
reports 0% packet loss.
