+++ crates/mvm-agentd/src/lib.rs
@@ -43,6 +43,10 @@ pub mod fs_rpc;
 /// normal wake) the guest reseeds its CSPRNG so two clones don't generate
 /// identical key material.
 pub mod genid;
+/// Post-mount guest environment setup shared by the two guest inits, so the
+/// legacy per-rootfs initrd and the universal-initramfs agent cannot drift.
+#[cfg(target_os = "linux")]
+pub mod guest_bootstrap;
 pub mod guest_mount;
 /// Shared in-guest network bring-up (eth0 up + DHCP + static fallback), used by
 /// both the builder VM init and the workload guest netinit.
